flaskr/requirements.txt
Flask
Werkzeug
python-doctr[torch]
python-dotenv
easyocr
google-genai
shapely
geopandas